Update integer types and conversions

+++ src/lib.rs
@@ -37,9 +37,9 @@ fn main() {
 
 extern crate log;
 
-pub static VERSION_MAJOR: uint = 0;
-pub static VERSION_MINOR: uint = 1;
-pub static VERSION_TINY: uint = 0;
+pub static VERSION_MAJOR: usize = 0;
+pub static VERSION_MINOR: usize = 1;
+pub static VERSION_TINY: usize = 0;
 pub static PRE_RELEASE: bool = true;
 
 /// Build version string.
